refactor(backend): extract tap/query mode into providers/generate_modes/tap.py
Behavior-preserving move of the inline tap/query single-`final` path out of
generate._event_stream into stream_tap: click resolution, planning, the
enter_scene/zoom_continue/fresh op split, the judged render loop on
deliberate-camera enters (including the fix/enter-loop-deadline cumulative
deadline, now derived from the threaded started + ingress_timeout_s), the
draft/main dual-task drive, grounding, and final assembly. Identical SSE
frames and order, same early-returns, error paths, flag gates, and spend
accounting. generate._event_stream now holds a thin async-for dispatch; deps
threaded explicitly as parameters (the injected-helper pattern the
edit/ascend/expand extractions set), provider calls left as module-attribute
access so existing monkeypatch tests keep biting. Module-level helpers stay in
generate.py (tests import them from there); only the tap-only
_ENTER_AS_TO_RENDER constant moved.
